requirements.txt
Flask
mysql-connector-python
pandas
openpyxl
plotly
kaleido
statsmodels

